Rename the SYSTEM module to MOTOR-DEFINITIONS
This prevents a naming conflict with Lumen's SYSTEM module.
