+++ bin/darwin/motor-definitions.lua
