+++ bin/linux/motor-definitions.lua
